+++ docs/CHANGELOG.md
@@ -20,8 +20,8 @@
   * Triplet losses now work with context.
   * `CheckExtraLabelsForWeightsAndMask` now (1) accepts weights and mask in any order; (2) normalize weights such that the sum is (non-masked) bathSize, 
     preserving the ratio. This way the mean will be 1.
-  * Losses with masks and weights fixed to properly normalize the weights for the non-masked out examples.
-    Also, now using MaskedReduceMean if there is a mask.
+  * Losses with masks and weights fixed so weights/mask can be given in any order.
+    Also, now using MaskedReduceMean if there is a mask, and all losses return a scalar.
 
 # v0.17.0: bitwise ops, triplet losses, new layers, fixes, and more.